requirements.txt
Django>=4.2.7
Pillow>=10.0.1
python-decouple>=3.8
gunicorn>=21.2.0
django-extensions>=3.2.3
whitenoise>=6.6.0
Werkzeug>=2.3.7
